Implemented add button, restructured code
