+++ compile.sh
