+++ package.json
@@ -53,10 +53,10 @@
     "watch:labextension": "jupyter labextension watch ."
   },
   "dependencies": {
-    "@jupyterlab/application": "^3.4.8",
+    "@jupyterlab/application": "^3.6.1",
     "@jupyterlab/apputils": "^3.4.8",
     "@jupyterlab/cells": "^3.4.7",
-    "@jupyterlab/notebook": "^3.4.7",
+    "@jupyterlab/notebook": "^3.6.1",
     "@jupyterlab/testutils": "^3.4.8",
     "@lumino/messaging": "^1.10.2",
     "@lumino/widgets": "^1.34.0"
