+++ pyproject.toml
@@ -4,6 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "2D-JupyterLab-TS"
+description = "2D JupyterLab in TypeScript"
 readme = "README.md"
 license = { file = "LICENSE" }
 requires-python = ">=3.7"
@@ -23,7 +24,7 @@ classifiers = [
 ]
 dependencies = [
 ]
-dynamic = ["version", "description", "authors", "urls", "keywords"]
+dynamic = ["version", "authors", "urls", "keywords"]
 
 [project.optional-dependencies]
 test = [
