+++ src/ContentFactoryWithFooterButton.ts
@@ -1,0 +1,26 @@
+import { CommandRegistry } from '@lumino/commands';
+import { ICellFooter, Cell } from '@jupyterlab/cells';
+import CellFooterWithButton from './CellFooterWithButton';
+import { NotebookPanel} from '@jupyterlab/notebook';
+
+ /**
+  * Extend the default implementation of an `IContentFactory`.
+  */
+export class ContentFactoryWithFooterButton extends NotebookPanel.ContentFactory {
+    constructor(
+      commands: CommandRegistry,
+      options?: Cell.ContentFactory.IOptions | undefined
+    ) {
+      super(options);
+      this.commands = commands;
+    }
+    /**
+     * Create a new cell header for the parent widget.
+     */
+    createCellFooter(): ICellFooter {
+      return new CellFooterWithButton(this.commands);
+    }
+    private readonly commands: CommandRegistry;
+}
+
+export default ContentFactoryWithFooterButton
